+++ src/Creeps/Roles/Room/Harvester.js
@@ -1,7 +1,7 @@
 // Harvester start here
 
 const HarvesterInfo = {
-    name:'drone',
+    name:'VIKING [harvester]',
     role:'Harvester',
     pattern:[WORK,CARRY],
     count:25,
@@ -9,8 +9,7 @@ const HarvesterInfo = {
     isForRoad:false,
     useBoost:null,
     moveParts:true,
-    skipCarry:true,
-    withoutMove:false,
+    skipCarry:false,
 }
 
 const moveParametersForGetResource = {
@@ -43,14 +42,25 @@ function HarvesterGetResource(creep) {
     if (droppedResource.length > 0 && droppedResource.Amount > creep.store.getCapacity()) {
         if (creep.pickup(droppedResource[0]) == ERR_NOT_IN_RANGE) creep.moveTo(droppedResource[0], {ignoreCreeps: true, heuristicWeight: 1.2, range: 1, reusePath: 50})
     } else {
-        if (!creep.memory.source) creep.memory.source = GetActiveSource(creep);
-        else {
-            if (Game.getObjectById(creep.memory.source).energy > 0) source = Game.getObjectById(creep.memory.source);
-            else creep.memory.source = GetActiveSource(creep)
-        }
+        const information = GetRoomInformation(creep.room.name);
 
-        if (source != false) {
-            if (creep.harvest(source) == ERR_NOT_IN_RANGE) creep.moveTo(source, {heuristicWeight: 1.2, range: 1, reusePath: 30});
+        const energyStructures = information.Room.StructuresWithResources.Energy;
+        const otherStructures = information.Room.StructuresWithResources.Other;
+
+        const droppedResources = creep.room.find(FIND_DROPPED_RESOURCES);
+
+        if (creep.room.controller.level < 3 && energyStructures.length > 0) {
+            if (creep.withdraw(energyStructures[0], RESOURCE_ENERGY) == ERR_NOT_IN_RANGE) creep.moveTo(energyStructures[0]);
+        } else {
+            if (!creep.memory.source) creep.memory.source = GetActiveSource(creep);
+            else {
+                if (Game.getObjectById(creep.memory.source).energy > 0) source = Game.getObjectById(creep.memory.source);
+                else creep.memory.source = GetActiveSource(creep)
+            }
+
+            if (source != false) {
+                if (creep.harvest(source) == ERR_NOT_IN_RANGE) creep.moveTo(source, {heuristicWeight: 1.2, range: 1, reusePath: 30});
+            }
         }
     }
 }
@@ -68,12 +78,7 @@ function HarvesterWork(creep) {
         else DoUpgrade(creep, {heuristicWeight: 1.2, range: 3, reusePath: 50});
     } else {
         if(info.Room.RefillStructures.All.length > 0 ) DoRefill(creep, info.Room.RefillStructures.All[0], {ignoreCreeps: true, heuristicWeight: 1.2, range: 1, reusePath: 50})
-        else if (info.Room.Other.ConstructionSites.Amount > 0) {
-            const target = Game.getObjectById(info.Room.Other.ConstructionSites.Array[0])
-            if (creep.build(target) == ERR_NOT_IN_RANGE) creep.moveTo(target, {heuristicWeight: 1.2, range: 3, reusePath: 50})
-        }
-        else DoUpgrade(creep, {heuristicWeight: 1.2, range: 3, reusePath: 50});
-
+        else DoBuild(creep)
     }
 }
 
@@ -88,7 +93,12 @@ const harvester = {
         if (creep.room.name == creep.memory.room) {
 
             if (creep.memory.trainRole && creep.memory.trainRole == 'Head') {
-
+                const target = Game.getObjectById(creep.memory.trainTarget);
+                if (!Game.getObjectById(target.memory.destinationId).structureType || Game.getObjectById(target.memory.destinationId).structureType != 'container') {
+                    if (target.pos.isNearTo(Game.getObjectById(target.memory.destinationId))) creep.memory.trainRole = '-';
+                } else {
+                    if (target.pos.isEqualTo(Game.getObjectById(target.memory.destinationId))) creep.memory.trainRole = '-';
+                }
             } else {
                 if (creep.store.getUsedCapacity() == 0) creep.memory.mode = 0;
                 else if (creep.store.getUsedCapacity() == creep.store.getCapacity()) creep.memory.mode = 1;